fix: rename module to httpcache and fix tests location
